Add files via upload
This script creates a list of users who are unable to obtain IP addresses from the DHCP server for unexplained reasons. This applies to wireless network clients based on the Cisco Catalyst 9800 WLC solution. For the proper functioning of the script, the following modules available in the python-wifi-modules repository are needed.
connecthandlerc9800
parsedtextfsmc9800

In addition, the following libraries are used:
time
textfsm
netmiko