Add files via upload
The script's task is to check whether Cisco access points are reachable via SSH from the specified location where we intend to log in.
This script will operate on a wireless network controller and access points from Cisco, running on the IOSXE software platform. It has been tested on software version 17.9.x, but it is unlikely to encounter issues on older software versions.